restructure of selectors package
