+++ soupsavvy/api.py
@@ -22,12 +22,8 @@ used for convenience and readability in some cases.
 
 from soupsavvy.selectors.base import SoupSelector
 from soupsavvy.selectors.combinators import SelectorList
-from soupsavvy.selectors.components import (
-    AndSelector,
-    HasSelector,
-    NotSelector,
-    XORSelector,
-)
+from soupsavvy.selectors.general import HasSelector
+from soupsavvy.selectors.logical import AndSelector, NotSelector, XORSelector
 
 
 def is_(
