+++ tests/soupsavvy/selectors/components/and_selector_test.py
@@ -3,7 +3,7 @@
 import pytest
 
 from soupsavvy.exceptions import NotSoupSelectorException, TagNotFoundException
-from soupsavvy.selectors.components import AndSelector
+from soupsavvy.selectors.logical import AndSelector
 from tests.soupsavvy.selectors.conftest import (
     MockClassMenuSelector,
     MockClassWidgetSelector,
